+++ STRM/src/main/cpp/strm/include/strm/impl/models/TfLiteYoloV5Classifier.hpp
@@ -27,7 +27,7 @@ class TfLiteYoloV5Classifier : public Classifier {
 
   const float* getClassConfidences(const int i) const override;
 
-  Rect reconstructBox(float x, float y, float w, float h, int imageWidth, int imageHeight) override;
+  Rect reconstructBox(float x, float y, float w, float h, float imageWidth, float imageHeight) override;
 
   TfLiteDelegate* delegate;
   std::unique_ptr<tflite::Interpreter> interpreter;
